# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://starwarslibrary-env.eba-pjw3m5ax.us-east-1.elasticbeanstalk.com","description":"Generated server url"}],"paths":{"/weapons/{id}":{"get":{"tags":["weapon-controller"],"operationId":"get_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WeaponOutDto"}}}}}},"put":{"tags":["weapon-controller"],"operationId":"put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeaponInDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WeaponOutDto"}}}}}},"delete":{"tags":["weapon-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"patch":{"tags":["weapon-controller"],"operationId":"patch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WeaponOutDto"}}}}}}},"/species/{id}":{"get":{"tags":["species-controller"],"operationId":"get_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SpeciesOutDto"}}}}}},"put":{"tags":["species-controller"],"operationId":"put_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SpeciesInDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SpeciesOutDto"}}}}}},"delete":{"tags":["species-controller"],"operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"patch":{"tags":["species-controller"],"operationId":"patch_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SpeciesOutDto"}}}}}}},"/ships/{id}":{"get":{"tags":["ship-controller"],"operationId":"get_5","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ShipOutDto"}}}}}},"put":{"tags":["ship-controller"],"operationId":"put_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShipInDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ShipOutDto"}}}}}},"delete":{"tags":["ship-controller"],"operationId":"delete_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"patch":{"tags":["ship-controller"],"operationId":"patch_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ShipOutDto"}}}}}}},"/planets/{id}":{"get":{"tags":["planet-controller"],"operationId":"get_7","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlanetOutDto"}}}}}},"put":{"tags":["planet-controller"],"operationId":"put_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlanetInDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlanetOutDto"}}}}}},"delete":{"tags":["planet-controller"],"operationId":"delete_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"patch":{"tags":["planet-controller"],"operationId":"patch_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlanetOutDto"}}}}}}},"/characters/{id}":{"get":{"tags":["character-controller"],"operationId":"get_9","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CharacterOutDto"}}}}}},"put":{"tags":["character-controller"],"operationId":"put_4","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CharacterInDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CharacterOutDto"}}}}}},"delete":{"tags":["character-controller"],"operationId":"delete_4","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"patch":{"tags":["character-controller"],"operationId":"patch_4","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CharacterOutDto"}}}}}}},"/weapons":{"get":{"tags":["weapon-controller"],"operationId":"get","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WeaponOutDto"}}}}}}},"post":{"tags":["weapon-controller"],"operationId":"post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeaponInDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WeaponOutDto"}}}}}}},"/species":{"get":{"tags":["species-controller"],"operationId":"get_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SpeciesOutDto"}}}}}}},"post":{"tags":["species-controller"],"operationId":"post_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SpeciesInDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SpeciesOutDto"}}}}}}},"/ships":{"get":{"tags":["ship-controller"],"operationId":"get_4","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ShipOutDto"}}}}}}},"post":{"tags":["ship-controller"],"operationId":"post_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShipInDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ShipOutDto"}}}}}}},"/planets":{"get":{"tags":["planet-controller"],"operationId":"get_6","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PlanetOutDto"}}}}}}},"post":{"tags":["planet-controller"],"operationId":"post_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlanetInDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlanetOutDto"}}}}}}},"/characters":{"get":{"tags":["character-controller"],"operationId":"get_8","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CharacterOutDto"}}}}}}},"post":{"tags":["character-controller"],"operationId":"post_4","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CharacterInDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CharacterOutDto"}}}}}}},"/health-check":{"get":{"tags":["health-check-controller"],"operationId":"test","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"WeaponInDto":{"type":"object","properties":{"name":{"type":"string"},"techLevel":{"type":"string"},"ammunition":{"type":"integer","format":"int32"},"weight":{"type":"number","format":"float"}}},"WeaponOutDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"techLevel":{"type":"string"},"ammunition":{"type":"integer","format":"int32"},"weight":{"type":"number","format":"float"}}},"SpeciesInDto":{"type":"object","properties":{"name":{"type":"string"},"nativePlanet":{"type":"integer","format":"int64"},"language":{"type":"string"},"meanHeight":{"type":"number","format":"float"},"meanLifespan":{"type":"number","format":"float"}}},"SpeciesOutDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"nativePlanet":{"type":"integer","format":"int64"},"language":{"type":"string"},"meanHeight":{"type":"number","format":"float"},"meanLifespan":{"type":"number","format":"float"}}},"ShipInDto":{"type":"object","properties":{"name":{"type":"string"},"model":{"type":"string"},"constructor":{"type":"string"},"length":{"type":"number","format":"float"}}},"ShipOutDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"model":{"type":"string"},"constructor":{"type":"string"},"length":{"type":"number","format":"float"}}},"PlanetInDto":{"type":"object","properties":{"name":{"type":"string"},"meanTemperature":{"type":"number","format":"float"},"terrain":{"type":"string"},"climate":{"type":"string"}}},"PlanetOutDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"meanTemperature":{"type":"number","format":"float"},"terrain":{"type":"string"},"climate":{"type":"string"}}},"CharacterInDto":{"type":"object","properties":{"name":{"type":"string"},"gender":{"type":"string"},"yearBirth":{"type":"string","format":"date"},"forceSensitive":{"type":"boolean"},"species":{"type":"integer","format":"int64"},"ship":{"type":"integer","format":"int64"},"weapon":{"type":"integer","format":"int64"},"homeWorld":{"type":"integer","format":"int64"}}},"CharacterOutDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"gender":{"type":"string"},"yearBirth":{"type":"string","format":"date"},"forceSensitive":{"type":"boolean"},"species":{"type":"integer","format":"int64"},"ship":{"type":"integer","format":"int64"},"weapon":{"type":"integer","format":"int64"},"homeWorld":{"type":"integer","format":"int64"}}}}}}